# config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BookReviewAPI",
    "version": "1.0.0",
    "description": "API for managing a book review platform"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/users/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User registered successfully" },
          "400": { "description": "Bad request" }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "Login a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful, returns JWT" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get user by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "User details" },
          "404": { "description": "User not found" }
        }
      },
      "put": {
        "summary": "Update user profile",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "email": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User updated" },
          "404": { "description": "User not found" }
        }
      },
      "delete": {
        "summary": "Delete user",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "User deleted" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/books": {
      "post": {
        "summary": "Create a book",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "author": { "type": "string" },
                  "genre": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Book created" },
          "400": { "description": "Bad request" }
        }
      },
      "get": {
        "summary": "List all books",
        "parameters": [
          { "name": "genre", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of books" }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "summary": "Get book by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Book details" },
          "404": { "description": "Book not found" }
        }
      },
      "put": {
        "summary": "Update book",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "author": { "type": "string" },
                  "genre": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Book updated" },
          "404": { "description": "Book not found" }
        }
      },
      "delete": {
        "summary": "Delete book",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Book deleted" },
          "404": { "description": "Book not found" }
        }
      }
    },
    "/reviews": {
      "post": {
        "summary": "Create a review",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bookId": { "type": "string" },
                  "rating": { "type": "number" },
                  "comment": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Review created" },
          "400": { "description": "Bad request" }
        }
      },
      "get": {
        "summary": "List all reviews",
        "parameters": [
          { "name": "bookId", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of reviews" }
        }
      }
    },
    "/reviews/{id}": {
      "get": {
        "summary": "Get review by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Review details" },
          "404": { "description": "Review not found" }
        }
      },
      "put": {
        "summary": "Update review",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rating": { "type": "number" },
                  "comment": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Review updated" },
          "404": { "description": "Review not found" }
        }
      },
      "delete": {
        "summary": "Delete review",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Review deleted" },
          "404": { "description": "Review not found" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}